pyproject.toml
[project]
name = "rhythmscape"
version = "0.1.0"
description = "A Critical Urban Diagnostic Assistant — rhythm, automobility, and embodied residue, mapped and narrated."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Dong-Wan Gimm", email = "dwgimm@gmail.com" },
]
keywords = [
    "critical-geography",
    "rhythmanalysis",
    "urban-studies",
    "automobility",
    "gis",
]

dependencies = [
    "httpx>=0.27",
    "tenacity>=8.2",
    "structlog>=24.1",
    "pandas>=2.2",
    "pyarrow>=16.0",
    "pyyaml>=6.0",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.urls]
Homepage = "https://github.com/thelevanter/rhythmscape"
Repository = "https://github.com/thelevanter/rhythmscape"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhythmscape"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
